feat: add createTypedData and refactor exports
