--- packages/eip712/src/payload/domain.ts
@@ -1,1 +1,0 @@
-// create domain
