+++ packages/eip712/src/payload/types.ts
@@ -6,5 +6,3 @@ export interface FlattenPayloadResponse {
   payload: JSONObject
   numMessages: number
 }
-
-export const payloadMsgFieldForIndex = (i: number) => `msg${i}`
